Refactoring some names.
