+++ __init__.py
@@ -1,7 +1,7 @@
 __title__ = 'eightchanAPI'
 __version__ = '0.9'
 __build__ = 0x020501
-__author__ = 'Dan Mercurio'
+__author__ = 'Dan Mercurio, Bartlomiej Mika'
 __license__ = 'LGPL'
 __copyright__ = 'Copyright 2015 Dan Mercurio'
